[#63] 친구 추가 플로우
